requirements.txt
python-dotenv
langchain-community
langchain-core
litellm
requests
transformers
torch
accelerate
fastapi
uvicorn
bitsandbytes
pydantic-settings 